improvements
fixed mosi byte transfer
changed from byte angle to 2bute pwm from master
changed freq to 100 hz o pca968 - working

+++ stm32_high_level/stm32_pio/include/stm_spi_data_structure.h
@@ -13,7 +13,9 @@
 
 extern struct Master_output
 {
-   uint8_t servo[18];
+   // uint8_t servo[18];
+   uint16_t servo[18];
+
    uint8_t flags;
 }master_output;
